backend/requirements.txt
flask
flask_cors
pandas
faiss-cpu
sentence-transformers
requests
